+++ client/package.json
@@ -22,7 +22,7 @@
     "graphql-request": "^3.7.0",
     "graphql-tag": "^2.12.6",
     "js-yaml": "^4.1.0",
-    "react": "^17.0.2",
+    "react": "^18.0.0",
     "react-dom": "^17.0.2",
     "react-hook-form": "^7.14.1",
     "react-query": "^3.34.0",
@@ -56,5 +56,8 @@
     "eslint-import-resolver-babel-module": "^5.3.1",
     "jest": "26.6.0",
     "tsconfig-paths": "^3.14.1"
+  },
+  "eslintConfig": {
+    "extends": "react-app"
   }
 }
