--- client/src/components/app/App.component.scss
@@ -1,11 +1,0 @@
-html, body {
-	background-color: white;
-	margin: 0;
-}
-
-.app-container {
-	min-height: 100vh;
-	display: flex;
-	flex-direction: column;
-	align-items: stretch;
-}
