--- client/src/components/footer/Footer.component.scss
@@ -1,21 +1,0 @@
-// .home-footer {
-//   margin-top: auto !important;
-//   position: fixed;
-//   bottom: 0;
-// }
-
-
-.home-footer {
-  background-color: #480a77;
-  // position: absolute;
-  // bottom: 0;
-  color: white;
-  margin-top: auto;
-  width: 100%;
-  display: flex;
-  justify-content: center;
-  padding: 20px;
-  font-size: 20px;
-  font-family: 'Work Sans';
-  font-weight: 300;
-}
