--- client/src/components/footer/Footer.component.tsx
@@ -1,13 +1,0 @@
-import React from 'react';
-
-import './Footer.component.scss'
-
-const Footer: React.FC = () => {
-  return (
-  <div className="home-footer">
-    Disclaimer: This resource is intended for purely research purposes. It should not be used for emergencies or medical or professional advice. 
-  </div>
-  )
-}
-
-export default Footer;
