--- client/src/components/header/Header.component.scss
@@ -1,22 +1,0 @@
-@import '../../common/styles/variables';
-
-header nav {
-	background-color: $primary-bg;
-	color: white;
-	position: relative;
-	height: 80px;
-
-	ul {
-		display: flex;
-		justify-content: flex-end;
-		list-style: none;
-		margin: 0 60px;
-		padding: 12px 0;
-
-		li {
-			margin-left: 60px;
-			font-family: 'Work Sans';
-			font-size: 18px;
-		}
-	}
-}
