+++ client/src/routes/public.tsx
@@ -1,35 +1,44 @@
 import { Suspense } from 'react';
 import { Navigate, Outlet } from 'react-router-dom';
 
-import Home from 'pages/home/home.page';
-import { Results } from 'pages/results';
+// import Home from 'pages/home/home.page';
+// import { Results } from 'pages/results';
 import { MainLayout } from 'components/Layout';
-// import { lazyImport } from '@/utils/lazyImport';
+import { lazyImport } from 'utils/lazyImport';
 
-// const { DiscussionsRoutes } = lazyImport(
-//   () => import('@/features/discussions'),
-//   'DiscussionsRoutes'
-// );
-// const { Dashboard } = lazyImport(() => import('@/features/misc'), 'Dashboard');
-// const { Profile } = lazyImport(() => import('@/features/users'), 'Profile');
-// const { Users } = lazyImport(() => import('@/features/users'), 'Users');
+
+const { Home } = lazyImport(() => import('pages/home'), 'Home');
+const { Results } = lazyImport(() => import('pages/results'), 'Results');
 
 const App = () => {
   return (
     <MainLayout>
-
+      <Suspense
+        fallback={
+          <div className="h-full w-full flex items-center justify-center">
+            {/* <Spinner size="xl" /> */}
+            spinner
+          </div>
+        }
+      >
+        <Outlet />
+      </Suspense>
     </MainLayout>
   );
 };
 
 export const publicRoutes = [
   {
-    path: '/',
+    path: '/app',
     element: <App />,
     children: [
-      { path: '/results', element: <Results /> },
-      { path: '/', element: <Home /> },
-      { path: '*', element: <Navigate to="." /> },
+      { path: '/app/results', element: <Results /> },
+      { path: '/app/home', element: <Home /> },
+      // { path: '*', element: <Navigate to="." /> },
     ],
   },
+  // {
+  //   path: '/home',
+  //   element: <Home />
+  // }
 ];
